requirements.txt
aiogram==3.3.0
openai==1.12.0
chromadb==0.4.22
PyMuPDF==1.23.14
python-dotenv==1.0.0
pytest==7.4.4
pytest-vcr==1.0.2
httpx==0.26.0
numpy<2.0.0
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
lxml==4.9.3 